Lista de la Renta de vehiculos,  vista de el filtro de vehículos y la obtención local de los datos necesarios para registrar el vehículos.

+++ app/src/main/java/ucne/edu/proyectofinalaplicada2/data/local/entities/RentaEntity.kt
@@ -10,5 +10,5 @@ class RentaEntity(
     val vehiculoId: Int?,
     val fechaRenta: String?,
     val fechaEntrega: String?,
-    val total: Int?
+    val total: Double?
 )
